backend/requirements.txt
Django==4.2.24
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.24.4
pydantic==2.5.2
reportlab==4.0.7
pytest==7.4.3
pytest-django==4.7.0
python-decouple==3.8
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
yfinance==0.2.28
requests==2.31.0
Pillow==10.1.0
django-allauth==0.57.0
